Renamed files SpatialWorkerConnection -> RealWorkerConnection, WorkerConnection -> SpatialWorkerConnection
